Validating
This file is the origin version of u_20171015. Going over this version again to validate the beamforming matrrix and making sure everything is correct. Change the way that users are attached to cluster in #281. Two users are within the same cluster as long as they have common serving BSs. For example, user 1 is served by BS [0 1], and user 2 is served by BS [1 2]. For the cluster containing BS [0 1], both user 1 and 2 are intra-cluster users. BS [0 1] need to work together to remove intra-user interference. This fact is not reflected before modification. Previously, users are within the same cluster only if their serving BSs are completely identical, which is not true for user-centric clustering.